test(assistant-stream): colocate the closed-args regression, warn on dropped deltas, add changeset
