+++ packages/assistant-stream/src/core/modules/assistant-stream.test.ts
@@ -303,52 +303,3 @@ describe("AssistantStreamController withParentId", () => {
     expect(text?.parentId).toBe("group-1");
   });
 });
-
-describe("DataStreamDecoder interleaved tool-call args (#5290)", () => {
-  const decodeLines = async (lines: string[]) => {
-    const bytes = new ReadableStream<Uint8Array>({
-      start(controller) {
-        const encoder = new TextEncoder();
-        for (const line of lines)
-          controller.enqueue(encoder.encode(line + "\n"));
-        controller.close();
-      },
-    });
-    const chunks: AssistantStreamChunk[] = [];
-    await bytes.pipeThrough(new DataStreamDecoder()).pipeTo(
-      new WritableStream({
-        write(chunk) {
-          chunks.push(chunk);
-        },
-      }),
-    );
-    return chunks;
-  };
-
-  it("does not crash when text arrives between tool-call begin and its args", async () => {
-    // Wire order from #5290: begin tool call, then an interleaved text delta
-    // (which closes the tool call's args stream), then an args delta for the
-    // same tool call id. Before the fix the final `c:` frame enqueued onto a
-    // closed controller and threw mid-decode.
-    const chunks = await decodeLines([
-      'b:{"toolCallId":"t1","toolName":"search"}',
-      '0:"progress text"',
-      'c:{"toolCallId":"t1","argsTextDelta":"{\\"q\\":1}"}',
-    ]);
-
-    // The interleaved text must survive, and the tool call part must exist.
-    expect(
-      chunks.some(
-        (c) => c.type === "text-delta" && c.textDelta === "progress text",
-      ),
-    ).toBe(true);
-    expect(
-      chunks.some(
-        (c) =>
-          c.type === "part-start" &&
-          c.part.type === "tool-call" &&
-          c.part.toolCallId === "t1",
-      ),
-    ).toBe(true);
-  });
-});
